requirements.txt
Jinja2==2.8
LEPL==5.1.3
MarkupSafe==0.23
PasteDeploy==1.5.2
PyYAML==3.11
Pygments==2.0.2
WebOb==1.7.1
WebTest==2.0.18
beautifulsoup4==4.3.2
boto==2.31.1
chaussette==1.2
cornice==1.0.0
coverage==3.7.1
docutils==0.12
gevent==1.1rc5
greenlet==0.4.9
iso8601==0.1.10
jsonpointer==1.9
mock==1.0.1
nose==1.3.4
mysql-connector==2.1.6
openpyxl==2.4.8
pbkdf2==1.3
py==1.4.26
pycrypto==2.6.1
pyramid-exclog==0.7
pyramid==1.7
pytest==2.6.4
python-coveralls==2.5.0
pytz==2014.10
repoze.lru==0.6
requests==2.7.0
setuptools==7.0
sh==1.11
simplejson==3.6.5
six==1.9.0
sphinxcontrib-httpdomain==1.4.0
translationstring==1.3
tzlocal==1.1.2
venusian==1.0
waitress==0.8.9
zope.deprecation==4.1.1
zope.interface==4.1.1
requests_mock==1.3.0
hypothesis
git+https://github.com/openprocurement/barbecue.git
git+https://github.com/openprocurement/python-json-patch.git
git+https://github.com/openprocurement/rfc6266.git

